refactor: restructure provider
To separate backend provider move github to its own folder
